# Optimize template tracker