--- a/modules/tracker/tt/include/visp3/tt/vpTemplateTracker.h
+++ b/modules/tracker/tt/include/visp3/tt/vpTemplateTracker.h
@@ -84,8 +84,7 @@ protected:
 
 #ifndef DOXYGEN_SHOULD_SKIP_THIS
   vpTemplateTrackerPointSuppMIInv *ptTemplateSupp;     // pour inverse et compo
-  vpTemplateTrackerPointSuppMIInv **ptTemplateSuppPyr; // pour inverse et
-                                                       // compo
+  vpTemplateTrackerPointSuppMIInv **ptTemplateSuppPyr; // pour inverse et compo
 #endif
 
   vpTemplateTrackerPointCompo *ptTemplateCompo;     // pour ESM