--- a/modules/tracker/tt/include/visp3/tt/vpTemplateTrackerSSDInverseCompositional.h
+++ b/modules/tracker/tt/include/visp3/tt/vpTemplateTrackerSSDInverseCompositional.h
@@ -59,21 +59,12 @@ protected:
   bool compoInitialised;
   vpMatrix HInv;
   vpMatrix HCompInverse;
-  bool useTemplateSelect; // use only the strong gradient pixels to compute
-                          // the Jabocian
-  // pour eval evolRMS
-  double evolRMS;
-  std::vector<double> x_pos;
-  std::vector<double> y_pos;
-  double threshold_RMS;
+  bool useTemplateSelect; // use only the strong gradient pixels to compute the Jabocian
 
 protected:
   void initHessienDesired(const vpImage<unsigned char> &I);
   void initCompInverse(const vpImage<unsigned char> &I);
   void trackNoPyr(const vpImage<unsigned char> &I);
-  void deletePosEvalRMS();
-  void computeEvalRMS(const vpColVector &p);
-  void initPosEvalRMS(const vpColVector &p);
 
 public:
   explicit vpTemplateTrackerSSDInverseCompositional(vpTemplateTrackerWarp *warp);
@@ -81,6 +72,5 @@ public:
   /*! Use only the strong gradient pixels to compute the Jabobian. By default
    * this feature is disabled. */
   void setUseTemplateSelect(bool b) { useTemplateSelect = b; }
-  void setThresholdRMS(double threshold) { threshold_RMS = threshold; }
 };
 #endif