… example/tracking/CMakeLists.txt
… example/tracking/templateTracker.cpp
… modules/tracker/tt/src/vpTemplateTracker.cpp
… modules/tracker/tt/src/warp/vpTemplateTrackerWarpHomography.cpp
… modules/tracker/tt/src/zncc/vpTemplateTrackerZNCCInverseCompositional.cpp
… modules/tracker/tt_mi/include/visp3/tt_mi/vpTemplateTrackerMI.h
… modules/tracker/tt_mi/include/visp3/tt_mi/vpTemplateTrackerMIBSpline.h
… modules/tracker/tt_mi/include/visp3/tt_mi/vpTemplateTrackerMIForwardAdditional.h
… modules/tracker/tt_mi/include/visp3/tt_mi/vpTemplateTrackerMIInverseCompositional.h
… modules/tracker/tt_mi/src/mi/vpTemplateTrackerMIESM.cpp
… modules/tracker/tt_mi/src/mi/vpTemplateTrackerMIForwardAdditional.cpp
… modules/tracker/tt_mi/src/mi/vpTemplateTrackerMIForwardCompositional.cpp
… modules/tracker/tt_mi/src/mi/vpTemplateTrackerMIInverseCompositional.cpp
… modules/tracker/tt_mi/src/tools/vpTemplateTrackerMIBSpline.cpp