This pull request relates issue #614

Optimize `vpTemplateTrackerMIInverseCompositional` to reduce the number of iterations used during optimization.

The following images highlights the improvements using `$ ./templateTracker -t 9 -c` example.

Results obtained with `evalRMS_eps=1e-3`: 
![test-10-3](https://user-images.githubusercontent.com/55136051/65892058-f02b7680-e3a5-11e9-87f4-522acf5c667f.png)

Results obtained with `evalRMS_eps=1e-4`: 
![test-10-4](https://user-images.githubusercontent.com/55136051/65892160-1c46f780-e3a6-11e9-90d8-265e5079f391.png)

Results obtained with `evalRMS_eps=1e-5`: 
![test-10-5](https://user-images.githubusercontent.com/55136051/65892190-28cb5000-e3a6-11e9-95d8-033b7baaaf42.png)

Results obtained with `evalRMS_eps=1e-6`: 
![test-10-6](https://user-images.githubusercontent.com/55136051/65892213-2f59c780-e3a6-11e9-9119-c3e75fc59a28.png)

We decided to retain  `evalRMS_eps=1e-4` which gives a good trade off.