@fspindle 2019-10-07
The following table shows performance improvement obtained on a MacBook pro that has 8 cores and 32GB RAM when the template tracker estimates an homography.

- The `Old Time` column gives the mean time in `ms` to process 100 images
- The `New Time` column gives the mean time in `ms` to process 100 images using changes introduced in this PR (especially thresholding the difference of residual between 2 successive iterations to end optimization loop)

| Tracker | Old Time  | New Time | Command line |
| --- | --- | --- | --- |
| SSD ESM |  763.1 | 31.4 | `./templateTracker -c -l 100 -r -t 0 -w 2` |
| SSD forward add. |  672.2 |  25.2 | `./templateTracker -c -l 100 -r -t 1 -w 1` |
| SSD forward comp. | 697.8  | 25.5  | `./templateTracker -c -l 100 -r -t 2 -w 1` |
| SSD inverse comp. | 20.7  |  18.7 | `./templateTracker -c -l 100 -r -t 3 -w 1` |
| ZNCC forward add. | 609.7  | 92.0 | `./templateTracker -c -l 100 -r -t 4 -w 1` |
| ZNCC inverse comp. | 27.2  | 24.8 | `./templateTracker -c -l 100 -r -t 5 -w 1` |
| ~~MI ESM~~|  NA | NA | `./templateTracker -c -l 100 -r -t 6 -w 2` |
| MI forward add. | 1917.4  | 159.3 | `./templateTracker -c -l 100 -r -t 7 -w 1` |
| MI forward comp. | 1583.1  | 119.7 | `./templateTracker -c -l 100 -r -t 8 -w 1` |
| MI inverse comp. | 260.1  | 63.4 | `./templateTracker -c -l 100 -r -t 9 -w 1` |

MI ESM produces an exception that needs to be investigated.